requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
PyPDF2>=3.0.1
fpdf2>=2.7.8
pandas>=2.2.0
openpyxl>=3.1.2